Edits to start screen scene, game over screen scene, and golden apples
Polished up the start screen scene. Added and fixed scripts to the game over screen, with functionality and remembering high scores and reflecting those back to the player. Added a golden apple that has a 1 in 5 chance to spawn in place of a regular apple, worth 300 points instead of 100 points.